WIP: Move implementation to external files
