+++ striot.cabal
@@ -21,6 +21,9 @@ library
                       Striot.LogicalOptimiser
                       Striot.Nodes
                       Striot.Nodes.Types
+                      Striot.Nodes.TCP
+                      Striot.Nodes.Kafka
+                      Striot.Nodes.MQTT
                       Striot.CompileIoT
                       Striot.StreamGraph
   -- other-modules:
@@ -49,6 +52,7 @@ library
                     , hw-kafka-client   >= 3.0.0
                     , mtl
                     , lens
+                    , envy
   hs-source-dirs:     src
   default-language:   Haskell2010
 
@@ -80,12 +84,16 @@ test-suite test-striot
                     , hw-kafka-client   >= 3.0.0
                     , mtl
                     , lens
+                    , envy
   other-modules:      Striot.CompileIoT
                       Striot.FunctionalIoTtypes
                       Striot.FunctionalProcessing
                       Striot.LogicalOptimiser
                       Striot.Nodes
                       Striot.Nodes.Types
+                      Striot.Nodes.TCP
+                      Striot.Nodes.Kafka
+                      Striot.Nodes.MQTT
                       Striot.StreamGraph
                       VizGraph
   default-language:   Haskell2010
